+++ requirements.txt
@@ -1,4 +1,4 @@
-alembic==1.11.3
+alembic==1.12.0
 annotated-types==0.5.0
 anyio==3.7.1
 asyncpg==0.28.0
@@ -8,9 +8,7 @@ click==8.1.7
 colorama==0.4.6
 dnspython==2.4.2
 email-validator==2.0.0.post2
-Faker==19.6.2
-fastapi==0.101.1
-fastapi-utils==0.2.1
+fastapi==0.103.2
 greenlet==2.0.2
 h11==0.14.0
 httpcore==0.17.3
