+++ scripts/populate_db.bat
@@ -1,0 +1,4 @@
+@echo off
+
+rem Populate db with initial data
+python core/fixtures/generate_game_genre_base.py
